+++ src/steps/constants.ts
@@ -1,6 +1,8 @@
 import {
   RelationshipClass,
+  RelationshipDirection,
   StepEntityMetadata,
+  StepMappedRelationshipMetadata,
   StepRelationshipMetadata,
 } from '@jupiterone/integration-sdk-core';
 
@@ -17,7 +19,7 @@ export const Steps = {
 };
 
 export const Entities: Record<
-  'WORKSPACE' | 'GROUP' | 'USER' | 'CLUSTER' | 'AWS_INSTANCE',
+  'WORKSPACE' | 'GROUP' | 'USER' | 'CLUSTER',
   StepEntityMetadata
 > = {
   WORKSPACE: {
@@ -40,19 +42,13 @@ export const Entities: Record<
     _type: 'databricks_cluster',
     _class: ['Cluster'],
   },
-  AWS_INSTANCE: {
-    resourceName: 'AWS Instance',
-    _type: 'aws_instance',
-    _class: ['Host'],
-  },
 };
 
 export const Relationships: Record<
   | 'WORKSPACE_HAS_GROUP'
   | 'WORKSPACE_HAS_CLUSTER'
   | 'GROUP_HAS_USER'
-  | 'USER_CREATED_CLUSTER'
-  | 'CLUSTER_IS_AWS_INSTANCE',
+  | 'USER_CREATED_CLUSTER',
   StepRelationshipMetadata
 > = {
   WORKSPACE_HAS_GROUP: {
@@ -79,10 +75,25 @@ export const Relationships: Record<
     _class: RelationshipClass.CREATED,
     targetType: Entities.CLUSTER._type,
   },
+};
+
+export const TargetEntities: Record<'AWS_INSTANCE', StepEntityMetadata> = {
+  AWS_INSTANCE: {
+    resourceName: 'AWS Instance',
+    _type: 'aws_instance',
+    _class: ['Host'],
+  },
+};
+
+export const MappedRelationships: Record<
+  'CLUSTER_IS_AWS_INSTANCE',
+  StepMappedRelationshipMetadata
+> = {
   CLUSTER_IS_AWS_INSTANCE: {
     _type: 'databricks_cluster_is_aws_instance',
     sourceType: Entities.CLUSTER._type,
     _class: RelationshipClass.IS,
-    targetType: Entities.AWS_INSTANCE._type,
+    targetType: TargetEntities.AWS_INSTANCE._type,
+    direction: RelationshipDirection.FORWARD,
   },
 };