+++ src/steps/workspace/index.test.ts
@@ -1,56 +1,28 @@
 import {
-  createMockStepExecutionContext,
+  executeStepWithDependencies,
   Recording,
 } from '@jupiterone/integration-sdk-testing';
-import { fetchWorkspaceDetails } from '.';
-import { integrationConfig } from '../../../test/config';
+import { buildStepTestConfigForStep } from '../../../test/config';
 import { setupDatabricksRecording } from '../../../test/recording';
-import { IntegrationConfig } from '../../config';
+import { Steps } from '../constants';
 
-describe('#fetchWorkspaceDetails', () => {
+describe('workspaceSteps', () => {
   let recording: Recording;
-
-  beforeEach(() => {
-    recording = setupDatabricksRecording({
-      directory: __dirname,
-      name: 'fetchWorkspaceDetails',
-    });
-  });
 
   afterEach(async () => {
     await recording.stop();
   });
 
-  test('should collect data', async () => {
-    const context = createMockStepExecutionContext<IntegrationConfig>({
-      instanceConfig: integrationConfig,
-    });
+  describe('#fetchWorkspaceDetails', () => {
+    test('should collect data', async () => {
+      recording = setupDatabricksRecording({
+        directory: __dirname,
+        name: 'fetchWorkspaceDetails',
+      });
 
-    await fetchWorkspaceDetails(context);
-
-    expect({
-      numCollectedEntities: context.jobState.collectedEntities.length,
-      collectedEntities: context.jobState.collectedEntities,
-      encounteredTypes: context.jobState.encounteredTypes,
-    }).toMatchSnapshot();
-
-    const workspaces = context.jobState.collectedEntities.filter((e) =>
-      e._type.includes('databricks_workspace'),
-    );
-    expect(workspaces.length).toBe(1);
-    expect(workspaces).toMatchGraphObjectSchema({
-      _class: ['Account'],
-      schema: {
-        additionalProperties: false,
-        properties: {
-          _rawData: {
-            type: 'array',
-            items: { type: 'object' },
-          },
-          _type: { const: 'databricks_workspace' },
-          name: { type: 'string' },
-        },
-      },
+      const stepConfig = buildStepTestConfigForStep(Steps.WORKSPACE);
+      const stepResult = await executeStepWithDependencies(stepConfig);
+      expect(stepResult).toMatchStepMetadata(stepConfig);
     });
   });
 });