+++ test/config.ts
@@ -1,5 +1,8 @@
+import { IntegrationInvocationConfig } from '@jupiterone/integration-sdk-core';
+import { StepTestConfig } from '@jupiterone/integration-sdk-testing';
 import * as dotenv from 'dotenv';
 import * as path from 'path';
+import { invocationConfig } from '../src';
 import { IntegrationConfig } from '../src/config';
 
 if (process.env.LOAD_ENV) {
@@ -16,3 +19,11 @@ export const integrationConfig: IntegrationConfig = {
   databricksAccessToken:
     process.env.DATABRICKS_ACCESS_TOKEN || DEFAULT_DATABRICKS_ACCESS_TOKEN,
 };
+
+export function buildStepTestConfigForStep(stepId: string): StepTestConfig {
+  return {
+    stepId,
+    instanceConfig: integrationConfig,
+    invocationConfig: invocationConfig as IntegrationInvocationConfig,
+  };
+}